refactor: unify event observer infra for integration tests
