+++ src/tests/integration/Starlights.Integration.Harness.Drivers/Elements/ManageProficienciesDriver.cs
@@ -14,7 +14,7 @@ public sealed class ManageProficienciesDriver : IDriver
 {
     private readonly IIntegrationHost _integration;
     private readonly ElementsDriverContext _driverContext;
-    private readonly ElementsEventObserverCollection _events;
+    private readonly EventObserverCollection _events;
     private readonly ManageProficienciesEndpointDriver _endpoints;
 
     public ManageProficienciesDriver(IIntegrationHost integration, ElementsDriverContext driverContext)
@@ -22,7 +22,7 @@ public sealed class ManageProficienciesDriver : IDriver
         _integration = integration;
         _driverContext = driverContext;
         _endpoints = _integration.GetDriver<ManageProficienciesEndpointDriver>();
-        _events = _integration.GetElementsEventObserverCollection();
+        _events = _integration.Events;
     }
 
     public async Task<Guid> CreateProficiencyAsync(CreateProperties properties)
