+++ src/tests/integration/Starlights.Integration.Harness/Eventing/EventObserverCollection.cs
@@ -1,7 +1,5 @@
 ﻿using System.Collections.Concurrent;
-using System.Diagnostics;
 using Microsoft.Extensions.Logging;
-using Moq;
 using Starlights.Integration.Extensions;
 using Starlights.Platform.Eventing;
 
@@ -19,10 +17,10 @@ public sealed class EventObserverCollection
         _cancellationToken = integration.CancellationToken;
     }
 
-    public EventObserverT<T> Event<T>() where T : IDomainEvent
+    public EventObserver<T> Event<T>() where T : IDomainEvent
     {
         var observer = _observers.GetOrAdd(typeof(T), _ => CreateObserver<T>());
-        if (observer is EventObserverT<T> typedObserver)
+        if (observer is EventObserver<T> typedObserver)
         {
             return typedObserver;
         }
@@ -36,7 +34,7 @@ public sealed class EventObserverCollection
         return Event<T>().HandleAsync(domainEvent);
     }
 
-    private EventObserverT<T> CreateObserver<T>() where T : IDomainEvent
+    private EventObserver<T> CreateObserver<T>() where T : IDomainEvent
     {
         return new(_cancellationToken);
     }
@@ -84,121 +82,5 @@ public sealed class EventObserverCollection
         {
             observer.ClearInvocations();
         }
-    }
-}
-
-internal interface IEventObserver
-{
-    void ClearInvocations();
-}
-
-public class EventObserverT<T> : IEventObserver where T : IDomainEvent
-{
-    private readonly CancellationToken _cancellationToken;
-    private readonly object _gate = new();
-    private readonly List<WaitRegistration> _waitRegistrations = [];
-
-    public Mock<IDomainEventHandler<T>> Mock { get; } = new();
-    public List<T> Events { get; } = [];
-
-    public EventObserverT(CancellationToken cancellationToken)
-    {
-        Mock.Setup(m => m.HandleAsync(It.IsAny<T>()))
-            .Callback<T>(OnEvent)
-            .Returns(Task.CompletedTask);
-
-        _cancellationToken = cancellationToken;
-    }
-
-    public void ClearInvocations()
-    {
-        Mock.Invocations.Clear();
-        Events.Clear();
-    }
-
-    public Task HandleAsync(T domainEvent)
-    {
-        return Mock.Object.HandleAsync(domainEvent);
-    }
-
-    public Task WaitForEvent(Predicate<T>? predicate = null, int count = 1)
-    {
-        if (count <= 0)
-        {
-            throw new ArgumentOutOfRangeException(nameof(count), count, "Count must be greater than zero.");
-        }
-
-        Trace.WriteLine($"[TRC] waiting for event {typeof(T).Name} .... current invocations: {Mock.Invocations.Count}");
-        lock (_gate)
-        {
-            var matchedExisting = Events.Count(evt => predicate == null || predicate(evt));
-            if (matchedExisting >= count)
-            {
-                return Task.CompletedTask;
-            }
-
-            var tcs = new TaskCompletionSource<bool>(TaskCreationOptions.RunContinuationsAsynchronously);
-            var registration = new WaitRegistration(predicate, count, matchedExisting, tcs);
-
-            registration.CancellationRegistration = _cancellationToken.Register(static state =>
-            {
-                var wait = (WaitRegistration)state!;
-                wait.TaskSource.TrySetCanceled();
-                Trace.WriteLine($"[TRC] waiting for event {typeof(T).Name} was cancelled due to test timeout");
-            }, registration);
-
-            tcs.Task.ContinueWith(
-                _ => RemoveWaitRegistration(registration),
-                CancellationToken.None,
-                TaskContinuationOptions.ExecuteSynchronously,
-                TaskScheduler.Default);
-
-            _waitRegistrations.Add(registration);
-            return tcs.Task;
-        }
-    }
-
-    private void OnEvent(T domainEvent)
-    {
-        lock (_gate)
-        {
-            Events.Add(domainEvent);
-
-            if (_waitRegistrations.Count == 0)
-            {
-                return;
-            }
-
-            foreach (var wait in _waitRegistrations)
-            {
-                if (wait.Predicate is null || wait.Predicate(domainEvent))
-                {
-                    wait.MatchedCount++;
-                    if (wait.MatchedCount >= wait.ExpectedCount)
-                    {
-                        wait.TaskSource.TrySetResult(true);
-                    }
-                }
-            }
-        }
-    }
-
-    private void RemoveWaitRegistration(WaitRegistration registration)
-    {
-        lock (_gate)
-        {
-            _waitRegistrations.Remove(registration);
-        }
-
-        registration.CancellationRegistration.Dispose();
-    }
-
-    private sealed class WaitRegistration(Predicate<T>? predicate, int expectedCount, int matchedCount, TaskCompletionSource<bool> taskSource)
-    {
-        public Predicate<T>? Predicate { get; } = predicate;
-        public int ExpectedCount { get; } = expectedCount;
-        public int MatchedCount { get; set; } = matchedCount;
-        public TaskCompletionSource<bool> TaskSource { get; } = taskSource;
-        public CancellationTokenRegistration CancellationRegistration { get; set; }
     }
 }
