+++ src/tests/integration/Starlights.Integration.Harness/Eventing/IEventObserver.cs
@@ -1,0 +1,6 @@
+﻿namespace Starlights.Integration.Eventing;
+
+internal interface IEventObserver
+{
+    void ClearInvocations();
+}
