+++ src/tests/integration/Starlights.Integration.Harness/Extensions/DomainEventExtensions.cs
@@ -3,14 +3,6 @@ using Starlights.Integration.Eventing;
 
 namespace Starlights.Integration.Extensions;
 
-public static class DomainEventExtensions
-{
-    public static ElementsEventObserverCollection GetElementsEventObserverCollection(this IIntegrationHost host)
-    {
-        return host.Services.GetRequiredService<ElementsEventObserverCollection>();
-    }
-}
-
 public static class EventExtensions
 {
     extension(IIntegrationHost host)
